[script.artwork.downloader] -v1.0.3
Downloads artwork for selected movie when multiple exist with same name

+++ script.artwork.downloader/service.py
@@ -13,17 +13,20 @@ from resources.lib.settings import _settings
 
 # starts update/sync
 def autostart():
+        xbmcaddon.Addon().setSetting(id="files_overwrite", value='false')
         settings = _settings()
         settings._get()
         addondir = xbmc.translatePath( utils.__addon__.getAddonInfo('profile') )
         tempdir = os.path.join(addondir, 'temp')
-        service_runtime  = "%s:00" % settings.service_runtime
+        service_runtime  = str('%.2d'%int(settings.service_runtime) + ':00')
         log('Service - Run at startup: %s'%settings.service_startup, xbmc.LOGNOTICE)        
         log('Service - Run as service: %s'%settings.service_enable, xbmc.LOGNOTICE)
-        log('Service - Time: %s:00'%service_runtime, xbmc.LOGNOTICE)
-        xbmcvfs.rmdir(tempdir)
+        log('Service - Time: %s'%service_runtime, xbmc.LOGNOTICE)
+        if xbmcvfs.exists(tempdir):
+            xbmcvfs.rmdir(tempdir)
+            log('Removing temp folder from previous run.')
         if settings.service_startup:
-            time.sleep(10)
+            time.sleep(15)
             xbmc.executebuiltin('XBMC.RunScript(script.artwork.downloader,silent=true)')
         if settings.service_enable:
             while (not xbmc.abortRequested):
